debugging data types

+++ app.py
@@ -8,7 +8,7 @@ import scipy as sp
 vehicles_df = pd.read_csv('vehicles_us.csv')
 vehicles_df['price'] = vehicles_df['price'].astype('float')
 vehicles_df['date_posted'] = pd.to_datetime(vehicles_df['date_posted'], format='%Y-%m-%d')
-
+vehicles_df['days_listed'] = vehicles_df['days_listed'].astype('float')
